complete bid done

+++ app/Http/Controllers/Superadmin/BannerController.php
@@ -57,11 +57,30 @@ class BannerController extends Controller
      */
     public function store(Request $request)
     {
+        // image one 
         $imageName = $request->file('image');
         $image = $imageName->getClientOriginalName();
         $directory = 'images/banner/';
         $imgUrl = $directory.$image;
         $imageName->move($directory,$image);
+        // image two
+        $imageTwo = $request->file('image_two');
+        $imagetwo = $imageTwo->getClientOriginalName();
+        $directorytwo = 'images/banner/';
+        $imgUrlTwo = $directorytwo.$imagetwo;
+        $imageTwo->move($directorytwo,$imagetwo);
+        //image three 
+        $imageThree = $request->file('image_three');
+        $imagethree = $imageThree->getClientOriginalName();
+        $directorythree = 'images/banner/';
+        $imgUrlThree = $directorythree.$imagethree;
+        $imageThree->move($directorythree,$imagethree);
+        //image four 
+        $imageFour = $request->file('image_four');
+        $imagefour = $imageFour->getClientOriginalName();
+        $directoryfour = 'images/banner/';
+        $imgUrlFour = $directoryfour.$imagefour;
+        $imageFour->move($directoryfour,$imagefour);
 
 
 
@@ -69,6 +88,9 @@ class BannerController extends Controller
         $banner->title = $request->title;
         $banner->description = $request->description;
         $banner->image = $imgUrl;
+        $banner->image_two = $imgUrlTwo;
+        $banner->image_three = $imgUrlThree;
+        $banner->image_four = $imgUrlFour;
         $banner->active = $request->active;
         $done = $banner->save();
 
@@ -119,23 +141,55 @@ class BannerController extends Controller
     public function update(Request $request, $id)
     {
         $check_image= $request->file('image');
-        if($check_image){
+        $check_imagetwo= $request->file('image_two');
+        $check_imagethree= $request->file('image_three');
+        $check_imagefour= $request->file('image_four');
+        if($check_image && $check_imagetwo && $check_imagethree && $check_imagefour){
             $oldImage = $request->oldimage;
-            if(file_exists($oldImage)){
+            $oldImageTwo = $request->oldimagetwo;
+            $oldImageThree = $request->oldimagethree;
+            $oldImageFour = $request->oldimagefour;
+            if(file_exists($oldImage && $oldImageTwo && $oldImageThree && $oldImageFour)){
                 unlink($oldImage);
+                unlink($oldImageTwo);
+                unlink($oldImageThree);
+                unlink($oldImageFour);
             }
+
             $banner = Banner::find($id);
+             // image one 
             $imageName = $request->file('image');
             $image = $imageName->getClientOriginalName();
             $directory = 'images/banner/';
             $imgUrl = $directory.$image;
             $imageName->move($directory,$image);
+            // image two
+            $imageTwo = $request->file('image_two');
+            $imagetwo = $imageTwo->getClientOriginalName();
+            $directorytwo = 'images/banner/';
+            $imgUrlTwo = $directorytwo.$imagetwo;
+            $imageTwo->move($directorytwo,$imagetwo);
+            //image three 
+            $imageThree = $request->file('image_three');
+            $imagethree = $imageThree->getClientOriginalName();
+            $directorythree = 'images/banner/';
+            $imgUrlThree = $directorythree.$imagethree;
+            $imageThree->move($directorythree,$imagethree);
+            //image four 
+            $imageFour = $request->file('image_four');
+            $imagefour = $imageFour->getClientOriginalName();
+            $directoryfour = 'images/banner/';
+            $imgUrlFour = $directoryfour.$imagefour;
+            $imageFour->move($directoryfour,$imagefour);
 
 
             $banner->title = $request->title;
             $banner->description = $request->description;
             $banner->active = $request->active;
             $banner->image = $imgUrl;
+            $banner->image_two = $imgUrlTwo;
+            $banner->image_three = $imgUrlThree;
+            $banner->image_four = $imgUrlFour;
             $done = $banner->save();
